added optional studyTitle to analysis
