--- src/main/java/com/odysseusinc/arachne/datanode/dto/converters/CommonStudyDTOToStudyConverter.java
@@ -1,52 +1,0 @@
-/*
- *
- * Copyright 2018 Odysseus Data Services, inc.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Company: Odysseus Data Services, Inc.
- * Product Owner/Architecture: Gregory Klebanov
- * Authors: Pavel Grafkin, Alexandr Ryabokon, Vitaly Koulakov, Anton Gackovka, Maria Pozhidaeva, Mikhail Mironov
- * Created: April 20, 2017
- *
- */
-
-package com.odysseusinc.arachne.datanode.dto.converters;
-
-import com.odysseusinc.arachne.commons.api.v1.dto.CommonStudyDTO;
-import com.odysseusinc.arachne.datanode.model.study.Study;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.core.convert.support.GenericConversionService;
-import org.springframework.stereotype.Component;
-
-/**
- * @author vkoulakov
- * @since 4/18/17.
- */
-@Component
-public class CommonStudyDTOToStudyConverter implements Converter<CommonStudyDTO, Study> {
-
-    @Autowired
-    public CommonStudyDTOToStudyConverter(GenericConversionService conversionService) {
-        conversionService.addConverter(this);
-    }
-
-    @Override
-    public Study convert(final CommonStudyDTO dto) {
-
-        Study study = new Study();
-        study.setId(dto.getId());
-        study.setTitle(dto.getName());
-        return study;
-    }
-}
