--- src/main/java/com/odysseusinc/arachne/datanode/repository/StudyRepository.java
@@ -1,33 +1,0 @@
-/*
- *
- * Copyright 2018 Odysseus Data Services, inc.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Company: Odysseus Data Services, Inc.
- * Product Owner/Architecture: Gregory Klebanov
- * Authors: Pavel Grafkin, Alexandr Ryabokon, Vitaly Koulakov, Anton Gackovka, Maria Pozhidaeva, Mikhail Mironov
- * Created: April 12, 2017
- *
- */
-
-package com.odysseusinc.arachne.datanode.repository;
-
-import com.odysseusinc.arachne.datanode.model.study.Study;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-/**
- * @author vkoulakov
- * @since 4/12/17.
- */
-public interface StudyRepository extends JpaRepository<Study, Long> {
-}
