+++ src/main/resources/db/migration/V20190711174410__remove_study_table.sql
@@ -1,0 +1,11 @@
+alter table analyses add column if not exists study_title varchar;
+
+update analyses set study_title = studies.title
+	from studies
+	where
+		studies.id = analyses.study_id and analyses.study_id is not null;
+
+alter table analyses drop constraint if exists fk_analyses_study_id;
+alter table analyses drop column if exists study_id;
+
+drop table studies;
